Add files via upload
Breast cancer is known to be the second leading cause of death among women in the world.
Breast cancer occurs when there is an abnormal growth of a mass of tissues of malignant cellsThese cells originate from the milk glands of the breast. The classification of these cells is determined
by the rate at which these unusual cells progress and the effect that they have on other
normal cells that can eventually affect the whole body . According to statistics of the World
Health Organization (WHO), breast cancer is the most common type of cancer occurring in
women globally, accounting for approximately 2.1 million new breast cancer cases [2]. In
2018, it is estimated that 627,000 women died from breast cancer–that is approximately 15% of
all cancer deaths among women. There are research studies and work in the literature on
detection and classification of breast cancers using some conventional methods [3–6] that
have not used any specific type of machine learning.